Refactor PR #74 to void PPI dependency.
Pull request #74 'show variable name as __VAR__' had a dependency on
the PPI module which made its inclusion in the master less atractive
for the users not using the feature. By refactoring the part of the
code that depended on PPI out in a separate module, only the users
that would like to use the feature needs to install PPI.

The code remaining in master (that does not depend on PPI) now implements
a plugin hook in the write_label() sub in Data::Printer::Object. An
option in the configuration file called 'caller_plugin' is used to
register a caller plugin. If such a plugin is registered,
write_label() will load the plugin and then call the plugin to do the
generation of the label string.

The refactored part of #74 that depends on PPI has been included in
the module Data::Printer::Plugin::Caller::PPI (currently on Github).